+++ leader/src/main/scala/io/shinto/amaterasu/leader/Kami.scala
