+++ leader/src/main/scala/io/shinto/amaterasu/leader/dsl/JobParser.scala
@@ -1,4 +1,4 @@
-package io.shinto.amaterasu.dsl
+package io.shinto.amaterasu.leader.dsl
 
 import java.util.concurrent.BlockingQueue
 
@@ -6,8 +6,9 @@ import com.fasterxml.jackson.databind.{JsonNode, ObjectMapper}
 import com.fasterxml.jackson.databind.node.ArrayNode
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import io.shinto.amaterasu.common.dataobjects.ActionData
-import io.shinto.amaterasu.common.execution.JobManager
-import io.shinto.amaterasu.common.execution.actions.{Action, ErrorAction, SequentialAction}
+import io.shinto.amaterasu.leader.execution.actions.{ErrorAction, SequentialAction}
+import io.shinto.amaterasu.leader.execution.JobManager
+import io.shinto.amaterasu.leader.execution.actions.{Action, ErrorAction, SequentialAction}
 import org.apache.curator.framework.CuratorFramework
 
 import scala.collection.JavaConverters._
