+++ leader/src/main/scala/io/shinto/amaterasu/leader/execution/JobManager.scala
@@ -1,13 +1,12 @@
-package io.shinto.amaterasu.common.execution
+package io.shinto.amaterasu.leader.execution
 
 import java.util.concurrent.BlockingQueue
 
-import io.shinto.amaterasu.enums.ActionStatus
-import org.apache.curator.framework.CuratorFramework
-import org.apache.zookeeper.CreateMode
 import io.shinto.amaterasu.common.dataobjects.ActionData
-import io.shinto.amaterasu.common.execution.actions.Action
 import io.shinto.amaterasu.common.logging.Logging
+import io.shinto.amaterasu.enums.ActionStatus
+import io.shinto.amaterasu.leader.execution.actions.Action
+import org.apache.curator.framework.CuratorFramework
 
 import scala.collection.concurrent.TrieMap
