+++ leader/src/main/scala/io/shinto/amaterasu/leader/package.scala
