+++ leader/src/test/scala/io/shinto/amaterasu/common/execution/JobRestoreTests.scala
@@ -4,13 +4,12 @@ import java.util.concurrent.LinkedBlockingQueue
 
 import io.shinto.amaterasu.common.dataobjects.ActionData
 import io.shinto.amaterasu.enums.ActionStatus
-
-import org.apache.curator.framework.{ CuratorFramework, CuratorFrameworkFactory }
+import io.shinto.amaterasu.leader.execution.{JobLoader, JobManager}
+import org.apache.curator.framework.{CuratorFramework, CuratorFrameworkFactory}
 import org.apache.curator.retry.ExponentialBackoffRetry
 import org.apache.curator.test.TestingServer
 import org.apache.zookeeper.CreateMode
-
-import org.scalatest.{ BeforeAndAfterEach, Matchers, FlatSpec }
+import org.scalatest.{BeforeAndAfterEach, FlatSpec, Matchers}
 
 import scala.io.Source
